requirements/local.txt
-r base.txt

django-debug-toolbar>=1.3.0
ipython>=3.1.0
py>=1.4.26
pytest>=2.7.0
pytest-django>=2.8.0
sqlparse>=0.1.14
pytz>=2015.2
six>=1.9.0
snowballstemmer>=1.2.0
Sphinx>=1.3.1
sphinx-rtd-theme>=0.1.7
Pygments>=2.0.2
Jinja2>=2.7.3
MarkupSafe>=0.23
docutils>=0.12
alabaster>=0.7.3
Babel>=1.3
